Audio memory usage tracking for more than 255 blocks

+++ teensy3/AudioStream.cpp
@@ -53,8 +53,8 @@ uint16_t AudioStream::memory_pool_first_mask;
 
 uint16_t AudioStream::cpu_cycles_total = 0;
 uint16_t AudioStream::cpu_cycles_total_max = 0;
-uint8_t AudioStream::memory_used = 0;
-uint8_t AudioStream::memory_used_max = 0;
+uint16_t AudioStream::memory_used = 0;
+uint16_t AudioStream::memory_used_max = 0;
